requirements.txt
aioquic==1.2.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.13
decorator==5.2.1
distro==1.9.0
exa-py==1.11.0
executing==2.2.0
firecrawl-py==1.15.0
Flask==3.1.0
gevent==24.11.1
google-auth==2.38.0
google-genai==1.9.0
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kaitaistruct==0.10
ldap3==2.9.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mitmproxy==11.0.2
mitmproxy-windows==0.10.7
mitmproxy_rs==0.10.7
msgpack==1.1.0
nest-asyncio==1.6.0
openai==1.70.0
packaging==24.2
parso==0.8.4
passlib==1.7.4
platformdirs==4.3.7
playwright==1.51.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psutil==7.0.0
publicsuffix2==2.20191221
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pydivert==2.1.0
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.10.1
pylsqpack==0.3.20
pyOpenSSL==24.3.0
pyparsing==3.2.0
pyperclip==1.9.0
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pywin32==310
pyzmq==26.4.0
requests==2.32.3
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
urwid==2.6.16
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
